pycommunicate_PY/requirements.txt
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.4
cryptography==42.0.8
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_daq==0.5.0
Flask==3.0.3
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
nest-asyncio==1.6.0
packaging==24.1
pillow==10.4.0
plotly==5.22.0
pycparser==2.22
pyserial==3.5
pywin32==306
pywinauto==0.6.8
requests==2.32.3
retrying==1.3.4
six==1.16.0
sympy==1.12.1
tenacity==8.4.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
